\name{tjGCluster2}
\alias{tjGCluster2}

\title{
 trajectory analysis
}

\description{
 trajectory analysis
}

\usage{
  tjGCluster2(m, from=NULL, to=NULL, cutoff=25);
}

\arguments{
   \item{m}{data.frame or matrix or igraph object}
   \item{from}{start vetex}
   \item{to}{end vetex}
   \item{cutoff}{it will be a branch if vertex number > cutoff}
}

\author{
Ying Hu \email{yhu@mail.nih.gov}

Chunhua Yan \email{yanch@mail.nih.gov}

Xiaoyu Liang \email{xiaoyu.liang@yale.edu}
}

\references{
##
}

\examples{
##
}

