\name{tree_list}
\alias{tree_list}

\title{
tree_list: Get the Tree Vertex List
}

\description{
Get the tree vertex list.
}

\usage{
tree_list(g, nodes);
}

\arguments{
   \item{g}{graph object}
   \item{nodes}{the spine skeleton vertices of graph}
}

\author{
Ying Hu \email{yhu@mail.nih.gov}

Chunhua Yan \email{yanch@mail.nih.gov}

Xiaoyu Liang \email{xiaoyu.liang@yale.edu}
}



